% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AJS_chad_doc.R
\docType{data}
\name{AJS_AmTiman}
\alias{AJS_AmTiman}
\title{Linelist data from an outbreak of Acute Jaundice Syndrome (AJS) in Chad 2016}
\format{
A data frame with 1477 rows and 122 columns:
\describe{
\item{variable1}{explanation of a variable}
\item{variable2}{more explanation of variables}
}
}
\source{
This data was collected as part of Doctors Without Borders (MSF) outbreak
response and the the Ministry of Health (MoH) in Chad has approved the use
of this data for training purposes.

The original dataset is AJS_AmTiman.xlsx.

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://www.gnu.org/licenses/gpl-3.0.html}{GPL3}
}
\usage{
AJS_AmTiman
}
\description{
Linelist data from a real outbreak of Hepatitis E virus (HEV) infection which
occurred in the town of Am Timan, Chad, between October 2016 and
April 2017.
}
\details{
This is a case-linelist which contains de-indentified individual level data
on clinical, laboratory and epidemiological information. It can be used
together with \link{AJS_AmTiman_population}, \link{Blocksshape} and
\link{Quartiersshape}.
Please note, that some data has been adapted in order to best achieve
training objectives. Also, the GPS coordinates included in the dataset
do not correspond to real cases identified during this outbreak.
They have been generated exclusively for training purposes.
}
\references{
{Spina A, Lenglet A, et al. (2017) A large outbreak of Hepatitis E virus genotype 1 infection in an urban setting in Chad likely linked to household level transmission factors, 2016-2017. PLoS ONE 12(11): e0188240} \doi{10.1371/journal.pone.0188240}
}
\keyword{datasets}
