% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acutejaundicesyndrome_outbreak_tcd_2016_doc.R
\docType{data}
\name{AJS_AmTiman_population}
\alias{AJS_AmTiman_population}
\title{Population data from an outbreak of Acute Jaundice Syndrome (AJS) in Chad 2016}
\format{
A data frame with 648 rows and 3 columns:
}
\source{
This data was collected as part of Doctors Without Borders (MSF) outbreak
response and the the Ministry of Health (MoH) in Chad has approved the use
of this data for training purposes.

The original dataset is 'AJS_AmTiman_population.xlsx'.
This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{CC by-NC-SA 4.0}
}
\usage{
AJS_AmTiman_population
}
\description{
Population data from a real outbreak of Hepatitis E virus (HEV) infection which
occurred in the town of Am Timan, Chad, between October 2016 and
April 2017.
}
\details{
This contains population counts for houses and individuals in each of
the housing blocks found in \link{Blocksshape} and can also be used together
with \link{AJS_AmTiman}.
}
\references{
{Spina A, Lenglet A, et al. (2017) A large outbreak of Hepatitis E virus genotype 1 infection in an urban setting in Chad likely linked to household level transmission factors, 2016-2017. PLoS ONE 12(11): e0188240} \doi{10.1371/journal.pone.0188240}
}
\keyword{datasets}
