% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_linelists_outbreak_sle_2014_doc.R
\docType{data}
\name{case_linelists_linelist_raw}
\alias{case_linelists_linelist_raw}
\title{Fictional case linelist data simulating properties from the West African Ebola outbreak of 2014}
\format{
A data frame with 6611 rows and 28 columns
}
\source{
This data was originally added to the {outbreaks} package, extensive
edits have been made for teaching purposes.
The data was originally simulated to match properties of the
West African Ebola outbreak of 2014-2015 (see {outbreaks} package documentation
for details).

The original dataset can be accessed using outbreaks::ebola_sim. Within this
package the original dataset is case_linelists_linelist_raw.xlsx.

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://opensource.org/license/MIT}{MIT}
}
\usage{
case_linelists_linelist_raw
}
\description{
Fictional case linelist data simulating properties from the West African
Ebola outbreak of 2014
}
\details{
This is fictional case-linelist which contains de-indentified individual level data
on clinical outcome, demographics and region. Same content as \link{linelist_raw},
repackaged under the \code{case_linelists_} naming for this bundle.
}
\references{
{WHO Ebola Response Team. 2014. Ebola virus disease in West Africa–the first 9 months of the epidemic and forward projections. The New England journal of medicine 371:1481–1495} \doi{10.1056/NEJMoa1411100}
}
\keyword{datasets}
