% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacttracing_outbreak_zzz_2020_doc.R
\docType{data}
\name{cases_clean}
\alias{cases_clean}
\alias{contacts_clean}
\alias{followups_clean}
\alias{relationships_clean}
\title{Fictional case data for a generic disease to demonstrate contact tracing}
\format{
A data frame with 56 rows and 86 columns

An object of class \code{data.frame} with 66 rows and 62 columns.

An object of class \code{data.frame} with 437 rows and 53 columns.

An object of class \code{data.frame} with 100 rows and 21 columns.
}
\source{
This data was originally sample contact tracing data from the
\href{https://worldhealthorganization.github.io/godata/}{Go.Data} platform

The original datasets are cases_clean.rds, contacts_clean.rds, followups_clean.rds,
relationships_clean.rds

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://www.gnu.org/licenses/gpl-3.0.html}{GPL3}
}
\usage{
cases_clean

contacts_clean

followups_clean

relationships_clean
}
\description{
Fictional case data for a generic disease to demonstrate contact tracing
}
\details{
This is fictional case-linelist which contains de-indentified individual level data,
based on go.data format.
}
\keyword{datasets}
