% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebola_outbreak_sle_2014_doc.R
\docType{data}
\name{cleaning_dict}
\alias{cleaning_dict}
\title{Example data dictionary for fictional Ebola Virus Disease data simulating properties from the
West African Ebola outbreak of 2014}
\format{
A data frame with 9 rows and 3 columns
}
\source{
This has been created as an example for use with  \link{linelist_raw},
which is cleaned using code in the \href{https://epirhandbook.com/en/new_pages/cleaning.html}{epiRhandbook cleaning chapter}.

The original dataset can be accessed using \code{get_data("linelist_raw")}

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://opensource.org/license/MIT}{MIT}
}
\usage{
cleaning_dict
}
\description{
Incomplete data dictionary used for demonstrating cleaning of raw version of fictional Ebola Virus
Disease data simulating properties from the West African Ebola outbreak of 2014
}
\details{
This is fictional data dictionary which contains recoding information.
}
\references{
{WHO Ebola Response Team. 2014. Ebola virus disease in West Africa–the first 9 months of the epidemic and forward projections. The New England journal of medicine 371:1481–1495} \doi{10.1056/NEJMoa1411100}
}
\keyword{datasets}
