% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_surveillance_zzz_2020.R
\docType{data}
\name{country_demographics}
\alias{country_demographics}
\alias{country_demographics_2}
\alias{deaths_countryA}
\alias{deaths_countryB}
\title{Fictional country demographic and mortality data to demonstrate standardisation}
\format{
Dataframes

An object of class \code{data.frame} with 18 rows and 3 columns.

An object of class \code{data.frame} with 18 rows and 4 columns.

An object of class \code{data.frame} with 18 rows and 4 columns.
}
\source{
This data was created for training purposes

The original dataset is country_demographics.csv

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{CC by-NC-SA 4.0}
}
\usage{
country_demographics

country_demographics_2

deaths_countryA

deaths_countryB
}
\description{
Fictional country demographic and mortality data to demonstrate standardisation
}
\details{
Fictional country demographic and mortality data to demonstrate standardisation
}
\keyword{datasets}
