% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_desc.R
\name{create_desc}
\alias{create_desc}
\title{Create Description of Variables for Roxygen Documentation}
\usage{
create_desc(
  dictionary,
  variable_name,
  type = NULL,
  description,
  copy_to_clipboard = TRUE
)
}
\arguments{
\item{dictionary}{A data frame listing at minimum the variable name, variable
type and a description column.}

\item{variable_name}{A column listing the names of variables}

\item{type}{A column listing the type of variable}

\item{description}{A column explaining what the variable contains}

\item{copy_to_clipboard}{if \code{TRUE} (default), the description text will be
copied to the user's clipboard with \code{\link[clipr:write_clip]{clipr::write_clip()}}. If \code{FALSE}, it
will be printed to the user's console.}
}
\value{
A character string which can be pasted in to the documentation file
}
\description{
This function uses a data frame (in a data dictionary format) to create a the
description section of 'format' for data set documentation with the Roxygen
package.
}
\examples{
\dontrun{
# Create description text for ... 
create_desc(data_dict, "variable_name", "type", "description")
}

}
