% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_outbreak_moz_2019_doc.R
\docType{data}
\name{district_weekly_count_data}
\alias{district_weekly_count_data}
\title{Fictional counts of malaria cases from Mozambique in 2019}
\format{
A data frame with 9628 rows and 4 columns
}
\source{
This data is fictional based on data from Mozambique.

The original dataset district_weekly_count_data.xlsx.

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{CC by-NC-SA 4.0}
}
\usage{
district_weekly_count_data
}
\description{
Fictional counts of malaria cases from Mozambique in 2019
}
\details{
This is fictional aggregate data with counts by date and region.
}
\keyword{datasets}
