% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebola_linelist_2014_doc.R
\docType{data}
\name{ebola_linelist_20141201}
\alias{ebola_linelist_20141201}
\title{Fictional Ebola linelist data from fSierra Leone, 2014}
\format{
An object of class \code{data.frame} with 539 rows and 40 columns.
}
\source{
Fictional data provided by Applied Epi for training purposes

This is version 1.0.

This data is English.

This data is licensed under \href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{CC by-NC-SA 4.0}
}
\usage{
ebola_linelist_20141201
}
\description{
Fictional linelist data based on the real emergence of Ebola in West Africa in 2014.
}
\details{
This is a case linelist which contains fictional individual level data from one country
on clinical, laboratory and epidemiological information.
It is intended for use as part of Applied Epi training courses,
alongside \link{mpox_aggregate_table} (cumulative case totals by date and country).
}
\keyword{datasets}
