% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_models_sle_2014_doc.R
\docType{data}
\name{epidemic_models_epinow_res}
\alias{epidemic_models_epinow_res}
\alias{epidemic_models_epinow_res_small}
\alias{epidemic_models_generation_time}
\alias{epidemic_models_incubation_period}
\title{Outputs from {epinow2}, fictional Ebola outbreak in Sierra Leone (2014) - epidemic models example}
\format{
A large list
}
\source{
The original dataset is epidemic_models_epinow_res.rds

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://opensource.org/license/MIT}{MIT}
}
\usage{
epidemic_models_epinow_res

epidemic_models_epinow_res_small

epidemic_models_generation_time

epidemic_models_incubation_period
}
\description{
Outputs from {epinow2} based on fictional Ebola Virus Disease data simulating
properties from the West African Ebola outbreak of 2014, repackaged for the
epidemic models example.
}
\details{
These outputs are produced using \link{linelist_cleaned_rds} with
\href{https://epiforecasts.io/EpiNow2/}{epinow2}. Same content as \link{epinow_res}.
}
\keyword{datasets}
