% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_models_sle_2014_doc.R
\docType{data}
\name{epidemic_models_estimate_samples}
\alias{epidemic_models_estimate_samples}
\title{Posterior estimate samples, fictional Ebola outbreak in Sierra Leone (2014) - epidemic models example}
\format{
A data frame with 1242724 rows and 8 columns
}
\source{
The original dataset is epidemic_models_estimate_samples.rds

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://opensource.org/license/MIT}{MIT}
}
\usage{
epidemic_models_estimate_samples
}
\description{
Posterior samples of the {epinow2} estimates, unbundled from
\link{epidemic_models_epinow_res} (\verb{$estimates$samples}) for the epidemic
models example.
}
\details{
See \link{epidemic_models_epinow_res} for the full model output this was
extracted from.
}
\keyword{datasets}
