% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_models_sle_2014_doc.R
\docType{data}
\name{epidemic_models_summary}
\alias{epidemic_models_summary}
\title{Model run summary, fictional Ebola outbreak in Sierra Leone (2014) - epidemic models example}
\format{
A data frame with 5 rows and 3 columns
}
\source{
The original dataset is epidemic_models_summary.rds

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://opensource.org/license/MIT}{MIT}
}
\usage{
epidemic_models_summary
}
\description{
Summary table, unbundled from \link{epidemic_models_epinow_res} (\verb{$summary})
for the epidemic models example.
}
\details{
See \link{epidemic_models_epinow_res} for the full model output this was
extracted from.
}
\keyword{datasets}
