% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebola_outbreak_sle_2014_doc.R
\docType{data}
\name{epinow_res}
\alias{epinow_res}
\alias{epinow_res_small}
\alias{generation_time}
\alias{incubation_period}
\title{Outputs from {epinow2} based on fictional Ebola Virus Disease data simulating properties from the West African Ebola outbreak of 2014}
\format{
A large list

An object of class \code{epinow} (inherits from \code{list}) of length 3.

An object of class \code{list} of length 5.

An object of class \code{list} of length 5.
}
\source{
This is direct output from use with \link{linelist_cleaned_rds}.

The original datasets are epinow_res.rds, epinow_res_small.rds,
generation_time.rds and incubation_period.rds

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://opensource.org/license/MIT}{MIT}
}
\usage{
epinow_res

epinow_res_small

generation_time

incubation_period
}
\description{
Outputs from {epinow2} based on fictional Ebola Virus Disease data simulating properties from the West African Ebola outbreak of 2014
}
\details{
These outputs are produced using \link{linelist_cleaned_rds} with \href{https://epiforecasts.io/EpiNow2/}{epinow2}
}
\keyword{datasets}
