% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_case_linelist_outbreak_sle_2014_doc.R
\docType{data}
\name{example_case_linelist_2020_10_08}
\alias{example_case_linelist_2020_10_08}
\title{Dated snapshot of fictional case linelist data simulating properties from the West African Ebola outbreak of 2014}
\format{
A data frame with 6611 rows and 28 columns
}
\source{
This data was originally added to the {outbreaks} package, extensive
edits have been made for teaching purposes. Within this package the
original dataset is example_case_linelist_2020-10-08.xlsx.

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://opensource.org/license/MIT}{MIT}
}
\usage{
example_case_linelist_2020_10_08
}
\description{
A dated snapshot (2020-10-08) of fictional case linelist data simulating
properties from the West African Ebola outbreak of 2014, prepared for an
iteration/loops example folder.
}
\details{
Same content as \link{linelist_raw} / \link{case_linelists_linelist_raw},
repackaged as one file in a dated series intended to demonstrate reading
in and combining multiple files.
}
\references{
{WHO Ebola Response Team. 2014. Ebola virus disease in West Africa–the first 9 months of the epidemic and forward projections. The New England journal of medicine 371:1481–1495} \doi{10.1056/NEJMoa1411100}
}
\keyword{datasets}
