% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_malaria_outbreak_moz_2019_doc.R
\docType{data}
\name{example_district_weekly_count_data}
\alias{example_district_weekly_count_data}
\title{Fictional counts of malaria cases from Mozambique (example copy)}
\format{
A data frame with 9628 rows and 4 columns
}
\source{
Within this package the original dataset is example_district_weekly_count_data.xlsx

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{CC by-NC-SA 4.0}
}
\usage{
example_district_weekly_count_data
}
\description{
Fictional counts of malaria cases from Mozambique, repackaged for an
iteration/loops example folder.
}
\details{
Same content as \link{district_weekly_count_data}.
}
\keyword{datasets}
