% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_influenza_outbreak_chn_2013_doc.R
\docType{data}
\name{example_fluH7N9_China_2013}
\alias{example_fluH7N9_China_2013}
\title{Linelist data from an outbreak of influenza A H7N9 in China 2013 (example copy)}
\format{
A data frame with 136 rows and 8 columns
}
\source{
Within this package the original dataset is example_fluH7N9_China_2013.csv

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://opensource.org/license/MIT}{MIT}
}
\usage{
example_fluH7N9_China_2013
}
\description{
Linelist data from an outbreak of influenza A H7N9 in China 2013, repackaged
for an iteration/loops example folder.
}
\details{
Same content as \link{fluH7N9_China_2013}.
}
\keyword{datasets}
