% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_case_linelist_outbreak_sle_2014_doc.R
\docType{data}
\name{example_other}
\alias{example_other}
\title{Other-hospital subset, fictional Ebola Virus Disease case linelist (West Africa, 2014)}
\format{
A data frame with 885 rows and 30 columns
}
\source{
Within this package the original dataset is \code{example_Other.csv}.

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://opensource.org/license/MIT}{MIT}
}
\usage{
example_other
}
\description{
Subset of the cleaned fictional Ebola Virus Disease case linelist for cases
reported at hospitals grouped as "Other", one of six per-hospital files
prepared for an iteration/loops example where multiple files are read in
and combined.
}
\details{
Row subset of \link{example_hospital_linelists} / \link{hospital_linelists}
filtered to cases reported at hospitals grouped as "Other". See also
\link{example_central_hospital}, \link{example_military_hospital},
\link{example_port_hospital}, \link{example_st_marks_maternity_hospital},
and \link{example_missing}.
}
\references{
{WHO Ebola Response Team. 2014. Ebola virus disease in West Africa–the first 9 months of the epidemic and forward projections. The New England journal of medicine 371:1481–1495} \doi{10.1056/NEJMoa1411100}
}
\keyword{datasets}
