% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feveriosis_surveillance_zzz_2025_doc.R
\docType{data}
\name{feveriosis_main_fr}
\alias{feveriosis_main_fr}
\title{Feveriosis surveillance data from Viraland from Jan 1 2025 to Aug 31 2026 \(French\)}
\format{
A dataframe with 427 rows and 13 columns.
}
\source{
Fictional data created by Applied Epi for training purposes

This is version 1.1

This dataset is French.

This dataset is licensed under \href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{CC by-NC-SA 4.0}
}
\usage{
feveriosis_main_fr
}
\description{
French version of the surveillance data for fictional disease feveriosis. Reported in fictional country Viraland in 2025 & 2026.
}
\details{
This is a linelist with fictional case-level data from the imaginary country Viraland, with
clinical and epidemiological information for cases of feveriosis (fictional disease).
The data is intended for use as part of the French version of the case study "Analysis of routine surveillance of feveriosis"
alongside the French \link{feveriosis_recent_fr} (linelist of feveriosis cases from Aug 31 - Sep 4, 2026 in Viraland).
}
\keyword{datasets}
