% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influenza_outbreak_chn_2013_doc.R
\docType{data}
\name{fluH7N9_China_2013}
\alias{fluH7N9_China_2013}
\title{Linelist data from an outbreak of influenza A H7N9 in China 2013.}
\format{
A data frame with 136 rows and 8 columns
}
\source{
This data was originally added to the {outbreaks} package, no edits
were made here.
The data was collated by Adam Kucharski et al. from ProMed, WHO, FluTrackers, news
reports and research articles. Transfer to R and documentation by Simon
Frost (sdwfrost@gmail.com).

The original dataset can be accessed using outbreaks::fluH7N9_China_2013

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://opensource.org/license/MIT}{MIT}
}
\usage{
fluH7N9_China_2013
}
\description{
Linelist data from a real outbreak of influenza A H7N9 in China 2013.
}
\details{
This is a case-linelist which contains de-indentified individual level data
on clinical outcome, demographics and region.
}
\references{
{A. Kucharski, H. Mills, A. Pinsent, C. Fraser, M. Van Kerkhove, C. A. Donnelly, and S. Riley. 2014. Distinguishing between reservoir exposure and human-to-human transmission for emerging pathogens using case onset data. PLOS Currents Outbreaks. Mar 7, edition 1} \doi{10.1371/currents.outbreaks.e1473d9bfc99d080ca242139a06c455f}
}
\keyword{datasets}
