% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fulton_doc.R
\docType{data}
\name{fulton_population}
\alias{fulton_population}
\title{Zip population data Fulton County (USA), 2020}
\format{
An object of class \code{data.frame} with 49 rows and 2 columns.
}
\source{
This population data was provided by Fulton County for training purposes.

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{CC by-NC-SA 4.0}
}
\usage{
fulton_population
}
\description{
Table of population size per zip code in Fulton County (USA), 2020
}
\details{
This is an aggregate dataset with population size per zip code in Fulton County (USA), 2020.
It can be used together with \link{fulton_data} or \link{fulton_data_clean}, which are raw and
clean versions of COVID-19 linelist data for Fulton County respectively.

Please note, that some data has been adapted in order to best achieve
training objectives.
}
\keyword{datasets}
