% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campylobacter_surveillance_deu_2001.R
\docType{data}
\name{germany_weather}
\alias{germany_weather}
\title{Weather data for Germany from 2001-2011}
\format{
A list
}
\source{
This data was obtained using the {ecmwfr} package as described in the
epiRhandbook chapter on timeseries. This data was originally collected with the
European Union Copernicus satelite and made available on their
\href{https://cds.climate.copernicus.eu/datasets/derived-era5-land-daily-statistics?tab=overview}{climate data store}

The original dataset is germany_weather.zip

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://www.gnu.org/licenses/gpl-3.0.html}{GPL3}
}
\usage{
germany_weather
}
\description{
Weather data for Germany from 2001-2011
}
\details{
Weather data for Germany from 2001-2011
}
\keyword{datasets}
