% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Retrieve and Load Datasets}
\usage{
get_data(
  name = NULL,
  language = NULL,
  group_identifier = NULL,
  usage = NULL,
  unique_identifier = NULL,
  assign = FALSE
)
}
\arguments{
\item{name}{(optional) A string specifying the name of the dataset to load.}

\item{language}{(optional) A string specifying the language of the dataset.
If used with \code{name}, \code{group_identifier}, or \code{unique_identifier}, the function
will filter datasets based on this language.}

\item{group_identifier}{(optional) A string identifying a group of related
datasets. If provided, the function will return all datasets within the group
unless further criteria are given.}

\item{usage}{(optional) A string representing where the datasets are used.
If provided, the function will return all datasets within the group unless
further criteria are given.}

\item{unique_identifier}{(optional) A string representing the unique identifier
for a dataset.}

\item{assign}{(optional) A logical value. If FALSE (default), the function
returns the dataset(s). If TRUE, the dataset(s) are assigned to the global
environment.}
}
\value{
If \code{assign = FALSE}, returns the dataset(s). If a single dataset is requested,
returns that dataset. If multiple datasets are requested, returns a named list
of datasets. If \code{assign = TRUE}, the dataset(s) are loaded into the global
environment and the function returns NULL invisibly.
}
\description{
The \code{get_data} function is designed to load specific datasets from the
\code{appliedepidata} package based on search criteria such as \code{name}, \code{language},
\code{group_identifier}, or \code{unique_identifier}. The function can either return
the dataset(s) or assign them into the global environment.

Depending on the type of dataset (e.g., regular data frame, spatial data (\code{sf}),
or list), appropriate post-processing is done to ensure proper handling and
loading.
}
\details{
The function operates as follows:
\enumerate{
\item \strong{Dataset Lookup:} It performs an internal lookup to identify datasets
matching the search criteria provided.
\item \strong{Loading Datasets:} For each dataset found:
\itemize{
\item It attempts to load the dataset using the \code{data()} function from
the \code{appliedepidata} package.
\item The dataset is loaded into a temporary environment to avoid global
environment clutter during the loading process.
}
\item \strong{Post-Processing Based on Object Type:} Once the dataset is loaded,
the function checks the type of object and handles it accordingly:
\itemize{
\item \strong{Data frames} are directly assigned to the global environment without
any further processing.
\item \strong{Phylo objects} (\code{phylo} class) are directly assigned to the global
environment without any further processing.
\item \strong{Stars objects} (\code{stars} class) are directly assigned to the global
environment without any further processing.
\item \strong{Epinow objects} (\code{epinow} class) are directly assigned to the global
environment without any further processing.
\item \strong{Spatial objects} (\code{sf} class) are processed using \code{sf::st_as_sf()} to
ensure correct handling of spatial data, and then assigned to the global
environment.
\item \strong{Lists} that are not data frames are split into individual objects and
assigned to the global environment as separate entities.
}
}

If any dataset cannot be found or loaded, the function will raise an error.
}
\examples{
\dontrun{
# Load a dataset by name and return it
my_data <- get_data(name = "AJS_AmTiman")

# Load a dataset by name and assign to global environment
get_data(name = "AJS_AmTiman", assign = TRUE)

# Load a dataset by name and language, assign custom name
linelist <- get_data(name = "AJS_AmTiman", language = "en")

# Load datasets by group identifier and return as list
multiple_datasets <- get_data(group_identifier = "acutejaundicesyndrome_outbreak_tcd_2016")

}

}
\seealso{
\code{appliedepidata::search_data()} for details on finding datasets and their criteria.
}
