% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_basics_afr_2020_doc.R
\docType{data}
\name{gis_africa_countries}
\alias{gis_africa_countries}
\title{Africa country boundaries (GIS basics example)}
\format{
A simple features (sf) data frame with 51 rows and 65 columns
}
\source{
\href{https://www.naturalearthdata.com/}{Natural Earth} Admin-0 countries
dataset. Natural Earth data is in the public domain and free to use without
restriction.

Within this package the original dataset is gis_africa_countries.json

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://www.naturalearthdata.com/about/terms-of-use/}{Public Domain}
}
\usage{
gis_africa_countries
}
\description{
Country-level administrative boundaries for the African continent, for the
GIS basics mapping example.
}
\details{
Natural Earth Admin-0 country boundaries, filtered to the African continent
(51 countries). Intended for use in choropleth/basemap exercises, e.g.
alongside \link{gis_covid_incidence}.
}
\keyword{datasets}
