% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_basics_sle_2020_doc.R
\docType{data}
\name{gis_covid_incidence}
\alias{gis_covid_incidence}
\title{Fictional cumulative COVID-19 incidence by country (GIS basics example)}
\format{
A data frame with 238 rows and 3 columns
}
\source{
Fictional data created by Applied Epi for training purposes.

Within this package the original dataset is gis_covid_incidence.csv

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{CC by-NC-SA 4.0}
}
\usage{
gis_covid_incidence
}
\description{
Fictional cumulative COVID-19 incidence by country, created for the GIS
basics choropleth mapping example.
}
\details{
This is fictional data created for teaching purposes, simulating a global
cumulative incidence table suitable for a world-map choropleth exercise.
}
\keyword{datasets}
