% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_basics_sle_2020_doc.R
\docType{data}
\name{gis_linelist_cleaned_with_adm3}
\alias{gis_linelist_cleaned_with_adm3}
\title{Cleaned fictional Ebola linelist joined with admin-3 spatial data (GIS basics example)}
\format{
A data frame with 1000 rows and 31 columns
}
\source{
This data is a cleaned version of \link{linelist_cleaned_rds}, which is cleaned
using code in the \href{https://epirhandbook.com/en/new_pages/cleaning.html}{epiRhandbook cleaning chapter}.

Within this package the original dataset is gis_linelist_cleaned_with_adm3.rds

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://opensource.org/license/MIT}{MIT}
}
\usage{
gis_linelist_cleaned_with_adm3
}
\description{
Subset of cleaned fictional Ebola Virus Disease case linelist data with
additional admin-3 spatial join fields, repackaged for the GIS basics
mapping example.
}
\details{
Same content as \link{linelist_cleaned_with_adm3}. Intended for use alongside
\link{gis_population_sle_admpop_adm3_2020} in choropleth mapping exercises.
}
\references{
{WHO Ebola Response Team. 2014. Ebola virus disease in West Africa–the first 9 months of the epidemic and forward projections. The New England journal of medicine 371:1481–1495} \doi{10.1056/NEJMoa1411100}
}
\keyword{datasets}
