% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_basics_sle_2020_doc.R
\docType{data}
\name{gis_population_sle_population_statistics_sierraleone_2020}
\alias{gis_population_sle_population_statistics_sierraleone_2020}
\title{Age/sex population proportions for Sierra Leone (GIS basics example)}
\format{
A data frame with 3 rows and 18 columns
}
\source{
This population data was downloaded from
\href{https://data.humdata.org/dataset/cod-ps-sle}{Humanitarian Data Exchange}.
No edits were made to this.

Within this package the original dataset is gis_population_sle_population_statistics_sierraleone_2020.xlsx

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://data.humdata.org/faqs/licenses}{CC by-IGO}
}
\usage{
gis_population_sle_population_statistics_sierraleone_2020
}
\description{
Age/sex population proportions for Sierra Leone, for the GIS basics
population pyramid example.
}
\details{
Same source as \link{sle_admpop_adm3_2020}. Intended for use in population
pyramid plotting exercises.
}
\keyword{datasets}
