% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebola_outbreak_sle_2014_doc.R
\docType{data}
\name{hospital_linelists}
\alias{hospital_linelists}
\title{Subset of cleaned fictional Ebola Virus Disease data simulating properties from the
West African Ebola outbreak of 2014}
\format{
A data frame with 454 rows and 30 columns
}
\source{
This data is a subset version of \link{linelist_cleaned_excel}, which is cleaned
using code in the \href{https://epirhandbook.com/en/new_pages/cleaning.html}{epiRhandbook cleaning chapter}.

The original dataset is hospital_linelist.xlsx

This is version 4.0.

This dataset is English.

This dataset is licensed under \href{https://opensource.org/license/MIT}{MIT}
}
\usage{
hospital_linelists
}
\description{
Subset of cleaned version of fictional Ebola Virus Disease data simulating
properties from the West African Ebola outbreak of 2014
}
\details{
This is fictional case-linelist which contains de-indentified individual level data
on clinical outcome, demographics and region.
}
\references{
{WHO Ebola Response Team. 2014. Ebola virus disease in West Africa–the first 9 months of the epidemic and forward projections. The New England journal of medicine 371:1481–1495} \doi{10.1056/NEJMoa1411100}
}
\keyword{datasets}
