% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert_other_zzz_2020_doc.R
\docType{data}
\name{likert_data}
\alias{likert_data}
\title{Fictional likert data for demonstration purposes}
\format{
Dataframes
}
\source{
This data was created for training purposes

The original dataset is likert_data.csv

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{CC by-NC-SA 4.0}
}
\usage{
likert_data
}
\description{
Fictional likert data for demonstration purposes
}
\details{
Fictional likert data for demonstration purposes
}
\keyword{datasets}
