% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebola_outbreak_sle_2014_doc.R
\docType{data}
\name{linelist_raw}
\alias{linelist_raw}
\title{Fictional Ebola Virus Disease data simulating properties from the West African Ebola outbreak of 2014}
\format{
A data frame with 136 rows and 8 columns
}
\source{
This data was originally added to the {outbreaks} package, extensive
edits have been made for teaching purposes.
The data was originally simulated to match properties of the
West African Ebola outbreak of 2014-2015 (see {outbreaks} package documentation
for details).

The original dataset can be accessed using outbreaks::ebola_sim

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://opensource.org/license/MIT}{MIT}
}
\usage{
linelist_raw
}
\description{
Fictional Ebola Virus Disease data simulating properties from the West African
Ebola outbreak of 2014
}
\details{
This is fictional case-linelist which contains de-indentified individual level data
on clinical outcome, demographics and region.
}
\references{
{WHO Ebola Response Team. 2014. Ebola virus disease in West Africa–the first 9 months of the epidemic and forward projections. The New England journal of medicine 371:1481–1495} \doi{10.1056/NEJMoa1411100}'
}
\keyword{datasets}
