% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list_datasets}
\alias{list_datasets}
\title{List Files in a Package}
\usage{
list_datasets(package_name)
}
\arguments{
\item{package_name}{A character string specifying the name of the R package for which to list files.}
}
\value{
A data frame with two columns:
\describe{
\item{\code{file_path}}{The full path to each file in the \code{data} or \code{inst} directories.}
\item{\code{directory}}{The directory (\code{data} or \code{inst}) where the file is located.}
}
}
\description{
This function lists all the files in the \code{data} and \code{inst} directories of a
specified R package.
}
\examples{
\dontrun{
# List all files in the data and inst directories of the "ggplot2" package
files <- list_datasets(package_name = "ggplot2")
print(files)
}

}
