% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_data.R
\name{lookup_dataset}
\alias{lookup_dataset}
\title{Lookup Dataset Information}
\usage{
lookup_dataset(
  name = NULL,
  language = NULL,
  group_identifier = NULL,
  unique_identifier = NULL
)
}
\arguments{
\item{name}{(optional) A string specifying the name of the dataset to search for.}

\item{language}{(optional) A string specifying the language of the dataset.
If used with \code{name}, \code{group_identifier}, or \code{unique_identifier}, the function
will filter datasets based on this language.}

\item{group_identifier}{(optional) A string identifying a group of related
datasets. If provided, the function will return all datasets within the group
unless further criteria are given.}

\item{unique_identifier}{(optional) A string representing the unique identifier
for a dataset.}
}
\value{
A \code{data.frame} containing dataset information if a matching dataset
is found. If no dataset is found or if multiple datasets are found and
more details are required (unless for a group), an error message is
returned with suggestions for further search refinement.
}
\description{
This function is an internal function used within the \code{get_data.R} and
\code{save_data.R} functions.
This function helps retrieve information (from \emph{tableoftables.xlsx}) about
datasets based on provided search criteria such as \code{name}, \code{language},
\code{group_identifier}, or \code{unique_identifier}. It returns the dataset that
matches the specified criteria or suggests alternatives if mismatches
are found.
}
\details{
The function performs searches in the following order:
\enumerate{
\item \strong{Searching by \code{name}:}
\itemize{
\item If a single dataset is found by \code{name}, the function checks if other
parameters like \code{language} or \code{group_identifier} match the dataset.
If they do not match, it suggests alternatives.
\item If multiple datasets are found, additional filtering is done based
on \code{language} or \code{group_identifier} (if specified).
}
\item \strong{Handling incorrect \code{language} or \code{group_identifier}:}
\itemize{
\item If a dataset is found by \code{name}, but the specified \code{language} or
\code{group_identifier} does not match, alternative datasets are suggested
based on the same \code{unique_identifier}.
}
\item \strong{Searching by \code{group_identifier}:}
\itemize{
\item If a \code{group_identifier} is provided, the function returns all datasets
within the group unless further refinement by \code{language} is necessary.
}
\item \strong{Searching by \code{unique_identifier}:}
\itemize{
\item If a \code{unique_identifier} is provided, the function will attempt to find
a dataset by that identifier and apply further filtering by \code{language}
or \code{name} if needed.
}
}
}
\examples{
\dontrun{
# Search by name
lookup_dataset(name = "AJS_AmTiman")

# Search by name and language
lookup_dataset(name = "AJS_AmTiman", language = "en")

# Search by group identifier and language
lookup_dataset(group_identifier = "acutejaundicesyndrome_outbreak_tcd_2016", language = "fr")

# Search by unique identifier
lookup_dataset(unique_identifier = "acutejaundicesyndrome_outbreak_tcd_2016_linelist_1_1_outbreak_2016")

# Search by name and group identifier
lookup_dataset(name = "AJS_AmTiman", group_identifier = "acutejaundicesyndrome_outbreak_tcd_2016")
}
}
