% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_outbreak_moz_2019_doc.R
\docType{data}
\name{malaria_facility_count_data}
\alias{malaria_facility_count_data}
\title{Fictional counts of malaria cases from Mozambique in 2019}
\format{
A data frame with 3038 rows and 10 columns
}
\source{
This data is fictional based on data from Mozambique. It has been adapted
from \link{district_weekly_count_data}.

The original dataset malaria_facility_count_Data.rds.

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{CC by-NC-SA 4.0}
}
\usage{
malaria_facility_count_data
}
\description{
Fictional counts of malaria cases from Mozambique in 2019
}
\details{
This is fictional aggregate data with counts by date, region and age group.
}
\keyword{datasets}
