% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_survey_zzz_2021_doc.R
\docType{data}
\name{mortality_survey_data}
\alias{mortality_survey_data}
\title{Fake data for a mortality survey}
\format{
Two data frames, 'Mortality Survey' with 424 rows and 16 columns, and
'hh_member' with 2874 rows and 24 columns:

\describe{
\item{Mortality Survey}{A data frame with 424 rows and 16 columns:
\describe{
\item{variable1}{Description of variable1}
\item{variable2}{Description of variable2}
# Add more variables as needed
}
}
\item{hh_member}{A data frame with 2874 rows and 24 columns:
\describe{
\item{variable1}{Description of variable1}
\item{variable2}{Description of variable2}
# Add more variables as needed
}
}
}
}
\source{
This data was generated with functions from the
\href{https://r4epi.github.io/epidict/index.html}{epidict} package which was created
as part of the R4Epis project funded by the Doctors Without Borders (MSF)
sappling nursery.

The original dataset is 'mortality_survey_data.xlsx'

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://www.gnu.org/licenses/gpl-3.0.html}{GPL3}
}
\usage{
mortality_survey_data
}
\description{
This is an example fake dataset of a mortality survey which was generated
using \href{https://r4epi.github.io/epidict/index.html}{epidict} package functions.
}
\details{
This is a fake survey dataset. This data comes in two parts:
a household level dataset ('Mortality Survey') and an individual
level dataset ('hh_member'). It can be used together with \link{mortality_survey_dict}.
Please note, that this data is entirely fictional created based on the
Doctors without Borders (MSF) ethical review board pre-approved survey
dictionaries. They have been generated here exclusively for training purposes.
}
\keyword{datasets}
