% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpox_outbreak_eue_2022_doc.R
\docType{data}
\name{mpox_aggregate_table}
\alias{mpox_aggregate_table}
\title{Fictional cumulative mpox case totals from five European countries, 2022}
\format{
An object of class \code{data.frame} with 101 rows and 3 columns.
}
\source{
Fictional data provided by ECDC EI Group for training purposes

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{CC by-NC-SA 4.0}
}
\usage{
mpox_aggregate_table
}
\description{
Fictional case counts based on the real emergence of Mpox in Europe in 2022.
}
\details{
This is a table with cumulative mpox case totals by date and country (country A - E).
It is intended for use as part of the Mpox case study, alongside \link{mpox_linelist} (case linelist
with fictional case-level information data from five countries).
}
\keyword{datasets}
