% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpox_outbreak_eue_2022_doc.R
\docType{data}
\name{mpox_linelist}
\alias{mpox_linelist}
\title{Fictional mpox linelist data from five European countries, 2022}
\format{
An object of class \code{data.frame} with 2000 rows and 13 columns.
}
\source{
Fictional data provided by ECDC EI Group for training purposes

This is version 1.0.

This data is English.

This data is licensed under \href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{CC by-NC-SA 4.0}
}
\usage{
mpox_linelist
}
\description{
Fictional linelist data based on the real emergence of mpox in Europe in 2022.
}
\details{
This is a case linelist which contains fictional individual level data from five countries
(country A - E) on clinical, laboratory and epidemiological information.
It is intended for use as part of the mpox case study,
alongside \link{mpox_aggregate_table} (cumulative case totals by date and country).
}
\keyword{datasets}
