% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidisease_tests_zzz_2024_doc.R
\docType{data}
\name{multidisease_tests}
\alias{multidisease_tests}
\title{Test results for notifiable disease surveillance cases in Feveria in 2024}
\format{
Dataframes
}
\source{
This data was created for training purposes

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{CC by-NC-SA 4.0}
}
\usage{
multidisease_tests
}
\description{
Test results for notifiable disease surveillance cases notified in fictional country Feveria in 2024, for malaria, dengue, typhoid, yellow fever, and cholera.
}
\details{
Fictional country surveillance data for the case study "Linking and analysing notifiable disease surveillance data and laboratory data in R"
}
\keyword{datasets}
