% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_survey_zzz_2021_doc.R
\docType{data}
\name{population}
\alias{population}
\title{Fake population data for a mortality survey}
\format{
A dataframe with 20 rows and 5 columns
}
\source{
This data was generated with functions from the
\href{https://r4epi.github.io/epidict/index.html}{epidict} package which was created
as part of the R4Epis project funded by the Doctors Without Borders (MSF)
sappling nursery.

The original dataset is 'population.xlsx'

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://www.gnu.org/licenses/gpl-3.0.html}{GPL3}
}
\usage{
population
}
\description{
This is an example fake population data for a mortality survey which was generated
using \href{https://r4epi.github.io/epidict/index.html}{epidict} package functions.
}
\details{
This is a fake population data for the \link{survey_data} dataset.
Please note, that this data is entirely fictional created based on the
Doctors without Borders (MSF) ethical review board pre-approved survey
dictionaries. They have been generated here exclusively for training purposes.
}
\keyword{datasets}
