% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shigella_surveillance_bel_2013_doc.R
\docType{data}
\name{sample_data_Shigella_tree}
\alias{sample_data_Shigella_tree}
\title{Linelist data to accompany next generation sequencing data of Shigella surveillance samples from Belgium 2013-2019}
\format{
A dataframe with 299 rows and 13 columns.
}
\source{
This data was posted on the \href{https://www.ebi.ac.uk/ena/browser/view/PRJEB40097?show=xrefs}{European Nucleotide Archive}

The original dataset is sample_data_Shigella_Tree.txt

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{CC by-NC-SA 4.0}
}
\usage{
sample_data_Shigella_tree
}
\description{
Linelist data to accompany next generation sequencing data of Shigella surveillance samples from Belgium 2013-2019
}
\details{
This linelist data to accompany sequencing data from routine surveillance samples collected in Belgium between 2013 and 2019.
It can be used together with \link{Shigella_tree}
}
\references{
{Fischer, N. et al. Genomic epidemiology of persistently circulating MDR Shigella sonnei strains associated with men who have sex with men (MSM) in Belgium (2013-19)} \doi{10.1093/jac/dkab377}
}
\keyword{datasets}
