% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_datasets.R
\name{search_datasets}
\alias{search_datasets}
\title{Search and Filter Datasets Available in appliedepidata}
\usage{
search_datasets()
}
\value{
A Shiny application object. This function will open the Shiny app in a web browser when executed.
}
\description{
Search and Filter Datasets Available in appliedepidata
}
\details{
Launches an interactive Shiny application that allows users to filter and
search a dataset.
Users can select a language from a dropdown menu and enter a search query to
filter the data displayed in a reactive table.
}
\examples{
\dontrun{
# Launch the interactive data search application
search_available_data()
}

}
