% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebola_outbreak_sle_2014_doc.R
\docType{data}
\name{sle_adm3}
\alias{sle_adm3}
\title{Shapefile at admin 3 level for Sierra Leone from Humanitarian Data Exchange}
\format{
A shapefile with 167 rows and 20 columns
}
\source{
This shapefile was downloaded from \href{https://data.humdata.org/dataset/cod-ab-sle}{Humanitarian Data Exchange}

This is the original dataset and was not edited (other than renaming)

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://data.humdata.org/faqs/licenses}{CC by-IGO}
}
\usage{
sle_adm3
}
\description{
Shapefile at admin 3 level for Sierra Leone from Humanitarian Data Exchange
}
\details{
This shapefile can be used together with \link{linelist_cleaned_with_adm3}.
Data collected from the government of Sierra Leone and cleaned and pcoded by UNOCHA and ITOS
Admin4 : Data provided by OCHA, attribute fields and geometries have been
checked by UNMEER IM (Ghana) Will be replaced if/when official Government data is provided
}
\keyword{datasets}
