% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebola_outbreak_sle_2014_doc.R
\docType{data}
\name{sle_hf}
\alias{sle_hf}
\title{Health facility points for Sierra Leone from OpenStreetMap}
\format{
A shapefile with 44 rows and 12 columns
}
\source{
This shapefile was downloaded from
\href{https://data.humdata.org/dataset/hotosm_sierra_leone_health_facilities}{Humanitarian Data Exchange}.
No edits were made to this.

This is the original dataset is sle_hf.shp

This is version 1.0.

This dataset is English.

This dataset is licensed under the \href{https://opendatacommons.org/licenses/odbl/1-0/}{Open Database License (ODbL)},
(c) OpenStreetMap contributors.
}
\usage{
sle_hf
}
\description{
Health facility points (hospitals, clinics, doctors, dentists, pharmacies) for
Sierra Leone, extracted from OpenStreetMap and published on the Humanitarian
Data Exchange.
}
\details{
This shapefile can be used together with \link{sle_adm3} and
\link{linelist_cleaned_with_adm3}. The Epidemiologist R Handbook's GIS basics
chapter joins each case to its nearest facility and draws catchment buffers
around the facilities. The export was made with the HOT Export Tool with the
filter \verb{healthcare IS NOT NULL OR amenity IN ('doctors','dentist','clinic', 'hospital','pharmacy')}.
}
\keyword{datasets}
