% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmonella_outbreak_de_1998_doc.R
\docType{data}
\name{stegentira_data}
\alias{stegentira_data}
\title{Linelist for fictionalized salmonella outbreak in Stegen, Germany}
\format{
A data frame with 291 rows and 20 columns
}
\source{
This case study was first designed by Alain Moren and Gilles Desve for EPIET. It is based on an investigation conducted by Anja Hauri, RKI, Berlin, 1998.
Data is fictional and was inspired by Nygren et al. Tick-borne encephalitis: acute clinical manifestations and severity in 581 cases from Germany, 2018-2020. Journal of Infection. 2023 Apr 1;86(4):369-75

This is version 1.0 (for the appliedepidata package. Earlier data versions used by EPIET are not considered)

This dataset is English.

This dataset is licensed under \href{https://www.gnu.org/licenses/gpl-3.0.html}{GPL3}
}
\usage{
stegentira_data
}
\description{
Linelist data for fictionalized salmonella outbreak that occured in a highschool
in Stegen, Germany (1998)
}
\details{
This is a fake case linelist with individual level data
on clinical and epidemiological information, with a focus on foods
eaten at a cafeteria (exposures)
}
\keyword{datasets}
