% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_survey_zzz_2021_doc.R
\docType{data}
\name{survey_dict}
\alias{survey_dict}
\title{Fake data dictionary for a mortality survey}
\format{
A dataframe with 46 rows and 14 columns
}
\source{
This data was generated with functions from the
\href{https://r4epi.github.io/epidict/index.html}{epidict} package which was created
as part of the R4Epis project funded by the Doctors Without Borders (MSF)
sappling nursery.

The original dataset is 'mortality_survey_dict.xlsx'

This is version 1.0.

This dataset is English.

This dataset is licensed under \href{https://www.gnu.org/licenses/gpl-3.0.html}{GPL3}
}
\usage{
survey_dict
}
\description{
This is an example fake data dictionary for a mortality survey which was generated
using \href{https://r4epi.github.io/epidict/index.html}{epidict} package functions.
}
\details{
This is a fake data dictionary for the \link{survey_data} dataset.
This data dictionary was originally in ODK excel format and using functions
from \href{https://r4epi.github.io/epidict/index.html}{epidict} was transformed in
to long format, so that it can be used for recoding using the
\href{https://www.repidemicsconsortium.org/matchmaker/}{matchmaker} package.
Please note, that this data is entirely fictional created based on the
Doctors without Borders (MSF) ethical review board pre-approved survey
dictionaries. They have been generated here exclusively for training purposes.
}
\keyword{datasets}
