% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidisease_surveillance_zzz_2024_doc.R
\docType{data}
\name{testy_multibolezni}
\alias{testy_multibolezni}
\title{Test results for notifiable disease surveillance cases in Feveria in 2024 \(Russian\)}
\format{
A dataframe with 1314 rows and 7 columns.
}
\source{
Fictional data created by Applied Epi for training purposes

This is version 1.0.

This dataset is Russian.

This dataset is licensed under \href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{CC by-NC-SA 4.0}
}
\usage{
testy_multibolezni
}
\description{
Russian version of the test results for notifiable disease surveillance cases notified in fictional country Feveria in 2024, for malaria, dengue, typhoid, yellow fever, and cholera.
}
\details{
This is the Russian version of a linelist with fictional test-level data from the imaginary country Feveria, with laboratory result information for suspected cases of malaria, dengue, typhoid, yellow fever, and cholera.
The data is intended for use as part of the case study "Linking and analyzing notifiable disease surveillance data and laboratory data in R"
alongside the Russian \link{uvedomleniya_multibolezni} (notification data in Feveria).
}
\keyword{datasets}
