% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidisease_surveillance_zzz_2024_doc.R
\docType{data}
\name{uvedomleniya_multibolezni}
\alias{uvedomleniya_multibolezni}
\title{Notifiable disease surveillance data from Feveria in 2024 \(Russian\)}
\format{
A dataframe with 987 rows and 12 columns.
}
\source{
Fictional data created by Applied Epi for training purposes

This is version 1.0.

This dataset is Russian.

This dataset is licensed under \href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{CC by-NC-SA 4.0}
}
\usage{
uvedomleniya_multibolezni
}
\description{
Russian version of the notifiable disease surveillance data for malaria, dengue, typhoid, yellow fever, and cholera. Reported in fictional country Feveria in 2024.
}
\details{
This is a linelist with fictional notification-level data from the imaginary country Feveria, with clinical and epidemiological information for
suspected cases of malaria, dengue, typhoid, yellow fever, and cholera.
The data is intended for use as part of the Russian version of the case study "Linking and analyzing notifiable disease surveillance data and laboratory data in R"
alongside the Russian \link{testy_multibolezni} (laboratory test results for notifications in Feveria).
}
\keyword{datasets}
