% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cholera_surveillance_zzz_2021.R
\docType{data}
\name{viraland_cholera_20210721_labs}
\alias{viraland_cholera_20210721_labs}
\title{Lab data with cholera test results for Viraland, 21 July 2021}
\format{
A dataframe with 45 rows and 6 columns.
}
\source{
Fictional data created by Applied Epi for training purposes

This is a derivate (subset) of the clean cholera surveillance data for Viraland, 21 July 2021.

This dataset is English.

This dataset is licensed under \href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{CC by-NC-SA 4.0}
}
\usage{
viraland_cholera_20210721_labs
}
\description{
Raw surveillance data for a cholera outbreak in fictional country Viraland in 2021, with data as of 21 July.
}
\details{
This is a linelist with fictional cholera test-level data from the imaginary country Viraland.
The data is intended for use alongside the intermediate qmd course and the intro to R course.
}
\keyword{datasets}
