% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cholera_surveillance_zzz_2021.R
\docType{data}
\name{viraland_cholera_20210721_linelist}
\alias{viraland_cholera_20210721_linelist}
\title{Clean cholera surveillance data for Viraland, 21 July 2021}
\format{
A dataframe with 1341 rows and 38 columns.
}
\source{
Fictional data created by Applied Epi for training purposes

This is the first version of data in this group; all others are derivates, including for 5th May (data version 3) and 17th June (data version 4).

This dataset is English.

This dataset is licensed under \href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{CC by-NC-SA 4.0}
}
\usage{
viraland_cholera_20210721_linelist
}
\description{
Clean surveillance data for a cholera outbreak in fictional country Viraland in 2021, with data as of 21 July.
}
\details{
This is a linelist with fictional case-level data from the imaginary country Viraland, with epidemiological, lab, and clinical information for
suspected cases of cholera.
The data is intended for use alongside the intermediate qmd course and the intro to R course.
}
\keyword{datasets}
