% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feveriosis_surveillance_zzz_2025_doc.R
\docType{data}
\name{feveriosis_recent}
\alias{feveriosis_recent}
\title{Recent feveriosis surveillance data from Feveria from Aug 31 2026 to Sep 4 2026}
\format{
A dataframe with 20 rows and 13 columns.
}
\source{
Fictional data created by Applied Epi for training purposes

This is version 1.1

This dataset is English.

This dataset is licensed under \href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{CC by-NC-SA 4.0}
}
\usage{
feveriosis_recent
}
\description{
Recent surveillance data for fictional disease feveriosis. Reported in fictional country Feveria from August 31, 2026 to September 4, 2026.
}
\details{
This is a linelist of a single week of case-level data of feveriosis (fictional data) from the imaginary country Feveria. The data is intended for use as part of the case study "Analysis of routine surveillance of feveriosis" alongside \link{feveriosis_main} (compiled linelist of feveriosis cases from Jan 1 2025 - Aug 31, 2026 in Feveria).
}
\keyword{datasets}
